Add support for global styles using styled-components injectGlobal. Documented caveats.
We don't want to use css loaders because it adds overhead. Maintaining loaders is annoying and it's just not real javascript/typescript.

CSS in JS is great because it allows us to simply use the capabilities of JS...
- modules
- functions
- so much more than SASS or LESS or whatever could ever provide

styled-components already abstracts away the generation of actual css files for performance.

WARNING:
- When editing global styles, all previous versions will persist until restarting the development server.
- Seems that styled-components is not flushing the 'injectGlobal' styles when creating new instances of ServerStyleSheet.
- This is not an issue in production.